+++ core/Updates/4.0.0-b1.php
@@ -49,12 +49,14 @@ class Updates_4_0_0_b1 extends PiwikUpdates
 
         $userModel = new Model();
         foreach ($userModel->getUsers(array()) as $user) {
-            $migrations[] = $this->migration->db->insert('user_token_auth', array(
-                'login' => $user['login'],
-                'description' => 'Created by Matomo 4 migration',
-                'password' => $userModel->hashTokenAuth($user['token_auth']),
-                'date_created' => Date::now()->getDatetime()
-            ));
+            if (!empty($user['token_auth'])) {
+                $migrations[] = $this->migration->db->insert('user_token_auth', array(
+                    'login' => $user['login'],
+                    'description' => 'Created by Matomo 4 migration',
+                    'password' => $userModel->hashTokenAuth($user['token_auth']),
+                    'date_created' => Date::now()->getDatetime()
+                ));
+            }
         }
 
         $migrations[] = $this->migration->db->dropColumn('user', 'token_auth');